requirements.txt
# =======================================
# Eidaah Backend - Dependencies
# =======================================

# FastAPI and Web Server
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# File Processing
pdfplumber==0.10.3
python-pptx==0.6.23

# AI - Groq API (OpenAI-compatible client)
openai==1.60.0

# Environment variables
python-dotenv==1.0.1

# Data validation
pydantic==2.9.0
